feat!: Create a server from Photon (as "Host")
BREAKING: we inject our own logic and override some Photon logic to do so.
This is logic for when a player clicks the "Create" button inside the Menu.
The Player will create a Lobby AS A CLIENT (so he is actually authed @ Menu scene Start).
(!) After creating, the server will show, but haven't tested joining or handled post-creation (eg: We don't join the game yet, nor do we close the modal UI).